Changing data type from long to int in line 18 for proper hahndling between functions.

+++ 04_functions/qsort.c
@@ -15,7 +15,7 @@ int main() {
 
 int test_qsort() {
     int array[] = {3, 5, 6, 9, 1, 10, 2, 0, 8, 1};
-    int SIZE = (long) sizeof(array) / sizeof(array[0]);
+    int SIZE = (int) sizeof(array) / sizeof(array[0]);
 
     printf("The initial array is:\n");
     print_array(array, SIZE);
